+++ lib/bwapi/client.rb
@@ -4,9 +4,12 @@ require 'bwapi/request'
 require 'bwapi/performance'
 
 require 'bwapi/client/admin'
+require 'bwapi/client/aggregates'
+require 'bwapi/client/author'
 require 'bwapi/client/brandwatch'
 require 'bwapi/client/client'
 require 'bwapi/client/command_center'
+require 'bwapi/client/dimensions'
 require 'bwapi/client/error_codes'
 require 'bwapi/client/filters'
 require 'bwapi/client/languages'
@@ -16,9 +19,9 @@ require 'bwapi/client/me'
 require 'bwapi/client/metrics'
 require 'bwapi/client/oauth'
 require 'bwapi/client/ping'
-require 'bwapi/client/sso'
 require 'bwapi/client/projects'
 require 'bwapi/client/query_validation'
+require 'bwapi/client/sso'
 require 'bwapi/client/test_search'
 require 'bwapi/client/user'
 require 'faraday'
@@ -32,9 +35,12 @@ module BWAPI
     include BWAPI::Performance
 
     include BWAPI::Client::Admin
+    include BWAPI::Client::Aggregates
+    include BWAPI::Client::Author
     include BWAPI::Client::Brandwatch
     include BWAPI::Client::Client
     include BWAPI::Client::CommandCenter
+    include BWAPI::Client::Dimensions
     include BWAPI::Client::ErrorCodes
     include BWAPI::Client::Filters
     include BWAPI::Client::Languages
@@ -45,8 +51,8 @@ module BWAPI
     include BWAPI::Client::OAuth
     include BWAPI::Client::Ping
     include BWAPI::Client::Projects
-    include BWAPI::Client::SSO
     include BWAPI::Client::QueryValidation
+    include BWAPI::Client::SSO
     include BWAPI::Client::TestSearch
     include BWAPI::Client::User
 
@@ -59,11 +65,19 @@ module BWAPI
       end
     end
 
+    # Check is access token has expired
+    #
+    # @return [Boolean] access token expiry status
+    def access_token_expired?
+      return true if @access_token.nil? || @access_token_expiry.nil?
+      seconds_until_access_token_expires <= 0
+    end
+
     # Check if user is authenicated
     #
     # @return [Boolean] Authenticated status
     def authenticated?
-      @access_token ? true : false
+      (@access_token && !access_token_expired?) ? true : false
     end
 
     # Check if user is a brandwatch-application-client type
@@ -108,6 +122,15 @@ module BWAPI
     def verify_ssl=(value)
       reset_connection
       @verify_ssl = value
+    end
+
+    private
+
+    # Returns the number of seconds until the access token expires
+    #
+    # @return [Integer] seconds until expiry
+    def seconds_until_access_token_expires
+      DateTime.parse(@access_token_expiry).to_time.to_i - Time.now.to_i
     end
   end
 end