+++ lib/bwapi/client/aggregates.rb
@@ -1,0 +1,13 @@
+module BWAPI
+  class Client
+    # Aggregates module for aggregates endpoints
+    module Aggregates
+      # Get the list of current aggregate dimensions
+      #
+      # @return [Array] All aggregates
+      def aggregates
+        get 'aggregates'
+      end
+    end
+  end
+end