+++ lib/bwapi/client/brandwatch/clients.rb
@@ -1,3 +1,4 @@
+require 'bwapi/client/brandwatch/clients/command_center'
 require 'bwapi/client/brandwatch/clients/users'
 require 'bwapi/client/brandwatch/clients/modules'
 
@@ -8,8 +9,6 @@ module BWAPI
       module Clients
         # Get specific clients
         #
-        # @note must be a super admin user
-        #
         # @param opts [Hash] options Hash of parameters
         # @options opts [String] nameContains Client name search (case insensitive)
         # @return [Hash] List of filtered clients
@@ -18,8 +17,6 @@ module BWAPI
         end
 
         # Create a new parent client
-        #
-        # @note must be a super admin user
         #
         # @param opts [Hash] options Hash of parameters
         # @option opts [String] address1 Address line one of the client
@@ -56,8 +53,6 @@ module BWAPI
 
         # Get specific client
         #
-        # @note must be a super admin user
-        #
         # @param client_id [Integer] Id of the client
         # @return [Hash] Specific client
         def brandwatch_get_client(client_id)
@@ -65,8 +60,6 @@ module BWAPI
         end
 
         # Update an existing parent client
-        #
-        # @note must be a super admin user
         #
         # @param client_id [Integer] Id of the client
         # @param opts [Hash] options Hash of parameters
@@ -102,6 +95,7 @@ module BWAPI
           put "brandwatch/clients/#{client_id}", opts
         end
 
+        include BWAPI::Client::Brandwatch::Clients::CommandCenter
         include BWAPI::Client::Brandwatch::Clients::Users
         include BWAPI::Client::Brandwatch::Clients::Modules
       end