+++ lib/bwapi/client/brandwatch/command_center.rb
@@ -1,0 +1,14 @@
+require 'bwapi/client/brandwatch/command_center/scene_types'
+require 'bwapi/client/brandwatch/command_center/themes'
+
+module BWAPI
+  class Client
+    module Brandwatch
+      # CommandCenter module for brandwatch/commandcenter endpoints
+      module CommandCenter
+        include BWAPI::Client::Brandwatch::CommandCenter::SceneTypes
+        include BWAPI::Client::Brandwatch::CommandCenter::Themes
+      end
+    end
+  end
+end