+++ lib/bwapi/client/command_center/category_colours.rb
@@ -1,0 +1,15 @@
+module BWAPI
+  class Client
+    module CommandCenter
+      # CategoryColours module for commandcenter/categoryColours endpoints
+      module CategoryColours
+        # Get colours defined for categories in current client
+        #
+        # TODO: Add parameters documentation
+        def all_category_colours
+          get 'commandcenter/categoryColours'
+        end
+      end
+    end
+  end
+end